feat: #1344; widget logic update; methods update;
